PWM implemented

+++ ARRIGONI_PIERGIORGIO.cydsn/InterruptRoutine.c
@@ -1,15 +1,3 @@
-/* ========================================
- *
- * Copyright YOUR COMPANY, THE YEAR
- * All Rights Reserved
- * UNPUBLISHED, LICENSED SOFTWARE.
- *
- * CONFIDENTIAL AND PROPRIETARY INFORMATION
- * WHICH IS THE PROPERTY OF your company.
- *
- * ========================================
-*/
-
 /**
 *   \file InterruptRoutine.h
 *   \brief Source file for the ISR triggered by the button
@@ -20,7 +8,7 @@
 
 CY_ISR(Button_ISR)
 {
-    extern int config;
+    extern uint8 config;
     
     config++;
     if(config == 8) //return to first configuration
